fix database conection problem
Change auto CRUD with springbood to manual SQL

+++ src/main/java/com/rodrigo/backend2java/model/AuthRequest.java
@@ -1,7 +1,10 @@
 package com.rodrigo.backend2java.model;
+
 import lombok.Data;
+
 @Data
 public class AuthRequest {
-    private long id;
+
+    private int id;
     private String senha;
 }
